requirements.txt
Django==4.2
django-ratelimit==4.1.0
djangorestframework==3.14.0
django-countries
Pillow==10.1.0
psycopg2==2.9.7
python-decouple==3.8
django-cors-headers==4.5.0

